Moved some functions to helper functions
